Remove references to file loader aborting.
Back when we were processing the files sequentially we could abort the
process by checking a flag if we should abort. Since we're now
processing the files in parallel, this kind of check no longer works so
references to abort should be removed. Maybe one day when we can
properly cancel promises, we can add the aborting again. For now, we
just return immediately if we try to load a folder while a folder is
already being loaded.