feat(spectator): prepare for spectating multiple streams at once.
The infrastructure is mainly complete but we should probably not enable
this feature until we can get the data stream logic and file-writing
logic in worker threads. Since this is all being done in the main thread
(which is fine when watching only one at a time) performance is probably
gonna tank when spectating many streams at once.