feat: disable auto-refresh for spectator page
The rationale behind this is, refreshing the page initiates a websocket
connection with the server. Most people won't need to spectate just
clicking through the pages so we don't want every average joe user
pinging the websocket server. So make sure we only connect to the server
on explicit refresh button clicks.