"Fixes" macOS child process issues.
On macOS, executing Slippi Dolphin as a child process from Node can
cause deadlocks when it comes time to render the game. This is most
notable with the Vulkan backend, but it's been exhibited with OpenGL
as well. Modern Macs require Vulkan (Metal) to perform at an
acceptable level, and newer versions of Slippi Dolphin now default to
Vulkan to ensure users are defaulted into a better playing experience.

To avoid these issues, this PR changes the Dolphin manager/instance
types to defer to the macOS `open` command. By doing this, we're
effectively uncoupling Slippi Dolphin's running entirely from the
Launcher, and thus removing whatever deadlocks were occurring.

As a result of not being able to own the child process, we can't
detect when Dolphin closes. On macOS this is really not a big deal,
though - the `open` command will route a user to whatever current
Dolphin instance (netplay || playback) is open. We avoid attaching
event listeners to the child process on macOS for this reason, though.

To handle cleanup on macOS, there's an addition to `main/utils` for
clearing out the temp comms files that were used on the last run of
the app. This happens at startup immediately now; if desired this
could be gated to macOS, but it shouldn't impact other OS's if left as
is.