Alternative fix for macOS hanging playback.
For whatever reason, there's some combination of flags that `spawn`
needs that are specific for macOS to not deadlock Dolphin as a child
process. Reading through the Node.js source, `execFile` does some work to
massage the options payload before ultimately calling `spawn` - so
rather than try to work around whatever oddities there are here, we can
just revert macOS to using execFile with a high maxBuffer setting.

There was a separate PR that piped things through the `open` command,
but that was more invasive and would cause issues with other parts of
the codebase. In comparison, this is much less invasive and doesn't
restrict the communication pattern used with Dolphin (I believe).