Check for updates to background, window sooner.
On macOS, prior to this change creating a new main window after having
closed it would take a few seconds. Forcing this check to the background
and showing the window sooner results in a quicker feeling load process.

I believe this should result in quicker appearance on other platforms as
well, but someone would need to test there.

This can be tested as of this commit by commenting out the `app.quit()`
call on `all-windows-closed`. Combined with some other PRs in the
pipeline it should work as expected.

(A good chunk of this patch was authored by Nikki; I just swapped some
window focus calls and tested on macOS)