Bump Electron to 13.1.8 (Universal macOS app).
This brings in a separate PR to bump Electron to a modern version, as
well as update the CI instructions to build a Universal app. One notable
change is that this disable the Metal feature flag on M1 devices; it
seems to cause freaked out rendering, and OpenGL on M1's is just a Metal
shim anyway, so it's no big deal.