deps: downgrade Electron from 16 to 13.
V13 is the last known version to have a working native file drop
functionality. At first glance it seemed like only the icon wasn't
showing when dragging a file, however it turns out on MacOS, dragging a
file ends up dragging the entire BrowserWindow. This makes the feature
pretty much unusable. And so unfortunately we'll have to downgrade by 3
major versions cuz the file drag and drop is a critical feature.