feat: add basic typesafety to slippi backend client GQL calls.
It's a little jank since you're essentially just redefining the type
each time you create a GQL call. So we're utilizing Pick in order to try
make it a bit more typesafe. But you can't easily use Pick on the child
nodes so if the children are also objects then we don't really have much
typesafety there. Not sure what we can do about that.