fix: replay file deletion sometimes throwing errors
Honestly not too sure why, but occasionally calling shell.trashItem
throws a "failed to create FileOperation instance" error. This often
happens when you first try to trash multiple items at once, before
trashing a single item. Once it successfully trashes an item, all future
multiple file trashes seems to work fine.

I have a hunch that this is caused by trying to delete multiple files at
the same time in the renderer thread. We attempt to solve this by
creating an IPC endpoing for deleting files, so that the file deletion
happens on the main thread instead. While we're at it we modify the
interface to be able to delete multiple files at once, rather than a
single file.

Also not entirely sure this actually fixes the problem but I haven't
seen another one of those errors after this change.