chore: disable type-checking during build.
We disable it since we already have a separate typecheck command
which is ran in CI as a part of the test process. Each CI run
essentially runs the typecheck twice - once for build, and once
for the typecheck command under the tests.

This reduces our build:renderer time from 60s -> 10s on my machine.